+++ guillotina/api/content.py
@@ -720,7 +720,11 @@ async def addable_types(context, request):
         for type_name in FACTORY_CACHE:
             types.append(type_name)
     app_settings["container_types"]
-    types = [item for item in types if item not in app_settings["container_types"]]
+    types = [
+        item
+        for item in types
+        if item not in app_settings["container_types"] and constrains.is_globally_allowed(item)
+    ]
     return types